Introduce tarerofs IO API
Previously, erofs-utils was integrated into overlaybd by
using the `mkfs.erofs` executable file. It's not optimial
since raw data needs to be dumped first and output data
needs to be write into overlaybd then.

This commit introduces IO API in tarerofs, that the data
stream of the:
- output image
- base image (if any)
- input tar file
can be controlled by the IO manager.

It should be noted that the mapfile is still handled by
reading and writing the temporary file 'upper.map', which
is an area for future improvement.

Signed-off-by: Hongzhen Luo <hongzhen@linux.alibaba.com>